Revert "[feature] implement Farcaster Frames for "Register to Vote" SWC actio…"
This reverts commit 39e7e349c32a943156b6a4661c175810c0f2569c.
