+++ package.json
@@ -31,7 +31,6 @@
     "seed": "npm run ts --transpile-only src/bin/seed/seedLocalDb.ts"
   },
   "dependencies": {
-    "@coinbase/onchainkit": "^0.12.0",
     "@graphql-tools/mock": "^9.0.0",
     "@hookform/resolvers": "^3.3.2",
     "@next/bundle-analyzer": "^14.1.0",
@@ -56,7 +55,6 @@
     "@radix-ui/react-tooltip": "^1.0.7",
     "@sentry/nextjs": "^7.91.0",
     "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@tanstack/react-query": "^5.29.0",
     "@tanstack/react-table": "^8.11.2",
     "@thirdweb-dev/auth": "^4.1.30",
     "@thirdweb-dev/engine": "^0.0.4",
@@ -115,7 +113,6 @@
     "swr": "^2.2.4",
     "tailwind-merge": "^2.2.0",
     "tailwindcss-animate": "^1.0.7",
-    "thirdweb": "^5.4.1",
     "twemoji": "^14.0.2",
     "twemoji-parser": "^14.0.0",
     "twitter-api-v2": "^1.15.2",
