+++ src/utils/server/nft/constants.ts
@@ -11,7 +11,7 @@ const CALL_REPRESENTATIVE_NFT_CONTRACT_ADDRESS = requiredEnv(
   'CALL_REPRESENTATIVE_NFT_CONTRACT_ADDRESS',
 )
 
-export const I_AM_A_VOTER_NFT_CONTRACT_ADDRESS = requiredEnv(
+const I_AM_A_VOTER_NFT_CONTRACT_ADDRESS = requiredEnv(
   process.env.I_AM_A_VOTER_NFT_CONTRACT_ADDRESS,
   'I_AM_A_VOTER_NFT_CONTRACT_ADDRESS',
 )
