+++ src/utils/server/thirdweb/thirdwebRPCClients.ts
@@ -1,7 +1,12 @@
 import { createPublicClient, http } from 'viem'
 import { base, mainnet } from 'viem/chains'
 
-import { THIRD_WEB_CLIENT_SECRET } from '@/utils/server/thirdweb/thirdwebClientSecret'
+import { requiredEnv } from '@/utils/shared/requiredEnv'
+
+const THIRD_WEB_CLIENT_SECRET = requiredEnv(
+  process.env.THIRD_WEB_CLIENT_SECRET,
+  'THIRD_WEB_CLIENT_SECRET',
+)
 
 export const thirdwebRPCClient = createPublicClient({
   chain: mainnet,
