--- src/utils/shared/neynarAPIKey.ts
@@ -1,1 +1,0 @@
-export const NEYNAR_API_KEY = 'NEYNAR_ONCHAIN_KIT' // Public API key.
